Moved request_id to strategy
