+++ lib/passport-wsfed-saml2/samlp.js
@@ -106,7 +106,7 @@ Samlp.prototype = {
     });
 
     var model = {
-      ID:               '_' + generateUniqueID(),
+      ID:               options.request_id,
       IssueInstant:     generateInstant(),
       Issuer:           options.realm,
       ProtocolBinding:  options.protocolBinding || BINDINGS.HTTP_POST,
@@ -400,15 +400,6 @@ Samlp.prototype = {
     });
   }
 };
-
-function generateUniqueID() {
-  var chars = "abcdef0123456789";
-  var uniqueID = "";
-  for (var i = 0; i < 20; i++) {
-    uniqueID += chars.substr(Math.floor((Math.random()*15)), 1);
-  }
-  return uniqueID;
-}
 
 function generateInstant() {
   var date = new Date();
