+++ lib/passport-wsfed-saml2/strategy.js
@@ -8,6 +8,7 @@ var wsfed     = require('./wsfederation');
 var samlp     = require('./samlp');
 var getReqUrl = require('./utils').getReqUrl;
 var EventEmitter = require('events');
+var utils     = require('./utils');
 
 var NullStateStore    = require('./state/null');
 var SessionStateStore = require('./state/session');
@@ -230,7 +231,9 @@ WsFedSaml2Strategy.prototype.authenticate = function (req, opts) {
     } else {
       // Initiate new samlp authentication request
       var authzParams = self.authorizationParams(opts);
-      
+      authzParams.request_id = '_' + utils.generateUniqueID()
+      meta.saml_request_id = authzParams.request_id;
+
       var sendRequestToIdp = function () {
         if (self.options.protocolBinding === 'urn:oasis:names:tc:SAML:2.0:bindings:HTTP-POST') {
           self._samlp.getSamlRequestForm(authzParams, function (err, form) {
