+++ lib/passport-wsfed-saml2/utils.js
@@ -4,6 +4,16 @@ function getReqUrl(req){
   return req.protocol + '://' + (req.get('x-forwarded-host') || req.get('host')) + req.originalUrl;
 }
 
-module.exports = {
-  getReqUrl : getReqUrl
+function generateUniqueID() {
+  var chars = "abcdef0123456789";
+  var uniqueID = "";
+  for (var i = 0; i < 20; i++) {
+    uniqueID += chars.substr(Math.floor((Math.random()*15)), 1);
+  }
+  return uniqueID;
 }
+
+module.exports = {
+  getReqUrl : getReqUrl,
+  generateUniqueID: generateUniqueID
+};
